better handle many edge cases
