+++ rex/exploit/technique.py
@@ -585,3 +585,30 @@ class Technique:
         if func_name in libc_obj.symbols_by_name:
             return libc_obj.symbols_by_name[func_name].rebased_addr
         return None
+
+    def _encode_cmd(self, cmd_str):
+        # command ending
+        if ord(b';') not in self.crash._bad_bytes:
+            cmd_str += b';'
+        elif ord(b'\n') not in self.crash._bad_bytes:
+            cmd_str += b'\n'
+        elif ord(b'\x00') not in self.crash._bad_bytes:
+            cmd_str += b'\x00'
+        elif ord(b'&') not in self.crash._bad_bytes:
+            cmd_str += b'&&1'
+        elif ord(b'|') not in self.crash._bad_bytes:
+            cmd_str += b'||1'
+
+        if ord(b' ') in self.crash._bad_bytes:
+            cmd_str = cmd_str.replace(b' ', b'$IFS')
+
+        # check shell coding
+        if any(x in self.crash._bad_bytes for x in b'$\\x\''):
+            return cmd_str
+        new_cmd_str = b''
+        for c in cmd_str:
+            if c not in self.crash._bad_bytes:
+                new_cmd_str += bytes([c])
+                continue
+            new_cmd_str += b"$'\\x%02x'" % c
+        return new_cmd_str
