+++ rex/exploit/techniques/ret2libc.py
@@ -42,21 +42,6 @@ class Ret2Libc(Technique):
 
         return True
 
-    def _encode_cmd(self, cmd_str):
-        if 0x20 in self.crash._bad_bytes:
-            cmd_str = cmd_str.replace(b' ', b'$IFS')
-
-        # check shell coding
-        if any(x in self.crash._bad_bytes for x in b'$\\x\''):
-            return cmd_str
-        new_cmd_str = b''
-        for c in cmd_str:
-            if c not in self.crash._bad_bytes:
-                new_cmd_str += bytes([c])
-                continue
-            new_cmd_str += b"$'\\x%2x'" % c
-        return new_cmd_str
-
     def _find_writable_region(self, data):
         segments = self.libc_rop.project.loader.main_object.segments
         for seg in segments:
@@ -158,8 +143,7 @@ class Ret2Libc(Technique):
         # if writing the cmd to memory is necessary, don't use null byte
         # or it will cause some troubles in ropchain generation
         # also, avoid blank space
-        cmd_str = cmd + b';'
-        cmd_str=  self._encode_cmd(cmd_str)
+        cmd_str = self._encode_cmd(cmd)
 
         try:
             if not cmd_addr:
