+++ rex/exploit/techniques/rop_to_system.py
@@ -80,7 +80,7 @@ class RopToSystem(Technique):
 
         # if writing the cmd to memory is necessary, don't use null byte
         # or it will cause some troubles in ropchain generation
-        cmd_str = cmd + b';'
+        cmd_str = self._encode_cmd(cmd)
 
         if not cmd_addr:
             # write out cmd_str
